Loop break for infinite loop problem
Either Pleroma or my handful of toots causes the status-fetching-with-pagination
to get stuck in an infinite loop. Seems to be a problem in the underlying
mastodon library rather than this code. This is an easier fix for now; to break
out if we fetched no statuses (i.e we ran out.)