Update README.md
Changed the example. First of all, now it fits better to the example output, second human subjects who try this tool out won't run in the problem that there account is freezed for several hours for there first test. And changed the example output to mach better with the real output.